plit dropdown into generic template and language-specific implementation. Move language logic to LanguageDropdown for better reusability.
